REFACTOR MigrationTest's getPackageProviders & getEnvironmentSetUp method to TestCase
